requirements.txt
fastapi
slowapi
uvicorn
openai
python-dotenv
pypdf